livekit-travel-assistant/rasa-bot/requirements.txt
rasa==3.6.20
rasa-sdk==3.6.2
pyyaml
apscheduler
fuzzywuzzy
geopy
python-Levenshtein
spacy==3.7.2
setuptools==65.5.1
pymongo==3.10.1
structlog==23.3.0
structlog-sentry==2.0.2
aiohttp==3.8.6
dnspython==2.3.0
protobuf==4.23.3
ruamel.yaml==0.17.21
scipy==1.10.1
tensorflow==2.12.0
tensorflow_hub==0.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-text==2.12.0
aiogram==2.25.2
pydantic==1.10.9